+++ pkg/addon/certpolicy/manifests/hubpermissions/role.yaml
