+++ pkg/addon/certpolicy/manifests/managedclusterchart/templates/_helpers.tpl
