+++ pkg/addon/certpolicy/manifests/managedclusterchart/templates/cluster_role.yaml
