+++ pkg/addon/certpolicy/manifests/managedclusterchart/templates/leader_election_role.yaml
