+++ pkg/addon/configpolicy/agent_addon.go
@@ -1,9 +1,8 @@
-package policy_framework
+package configpolicy
 
 import (
 	"context"
 	"embed"
-	"strings"
 
 	"github.com/openshift/library-go/pkg/assets"
 	"github.com/openshift/library-go/pkg/controller/controllercmd"
@@ -31,7 +30,7 @@ const (
 	// addOnAgentInstallationNamespace is the namespace on the managed cluster to install the helloworldhelm addon agent.
 	addOnAgentInstallationNamespace = "open-cluster-management-agent-addon"
 	defaultExampleImage             = "quay.io/open-cluster-management/helloworld-addon:latest"
-	addonName                       = "governance-policy-framework"
+	addonName                       = "config-policy-controller"
 )
 
 func init() {
@@ -39,15 +38,15 @@ func init() {
 }
 
 //go:embed manifests
-//go:embed manifests/managedcluster-chart
-//go:embed manifests/managedcluster-chart/templates/_helpers.tpl
+//go:embed manifests/managedclusterchart
+//go:embed manifests/managedclusterchart/templates/_helpers.tpl
 var FS embed.FS
 
 var agentPermissionFiles = []string{
 	// role with RBAC rules to access resources on hub
-	"manifests/hub-permissions/role.yaml",
+	"manifests/hubpermissions/role.yaml",
 	// rolebinding to bind the above role to a certain user group
-	"manifests/hub-permissions/rolebinding.yaml",
+	"manifests/hubpermissions/rolebinding.yaml",
 }
 
 func newRegistrationOption(kubeConfig *rest.Config, recorder events.Recorder, agentName string) *agent.RegistrationOption {
@@ -113,10 +112,9 @@ type global struct {
 }
 
 type userValues struct {
-	ClusterNamespace  string `json:"clusterNamespace"`
-	LogLevel          int32  `json:"logLevel"`
-	Global            global `json:"global"`
-	OnMulticlusterHub bool   `json:"onMulticlusterHub"`
+	ClusterNamespace string `json:"clusterNamespace"`
+	LogLevel         int32  `json:"logLevel"`
+	Global           global `json:"global"`
 }
 
 func getValues(cluster *clusterv1.ManagedCluster,
@@ -135,21 +133,6 @@ func getValues(cluster *clusterv1.ManagedCluster,
 				"NO_PROXY":    "",
 			},
 		},
-		OnMulticlusterHub: false,
-	}
-	// special case for local-cluster
-	if cluster.Name == "local-cluster" {
-		userValues.OnMulticlusterHub = true
-	}
-	if val, ok := addon.GetAnnotations()["addon.open-cluster-management.io/on-multicluster-hub"]; ok {
-		if strings.EqualFold(val, "true") {
-			userValues.OnMulticlusterHub = true
-		} else if strings.EqualFold(val, "false") {
-			// the special case can still be overridden by this annotation
-			userValues.OnMulticlusterHub = false
-		} else {
-			// TODO: should this log or return an error? The annotation should be true or false
-		}
 	}
 	return addonfactory.JsonStructToValues(userValues)
 }
@@ -160,7 +143,7 @@ func GetAgentAddon(controllerContext *controllercmd.ControllerContext) (agent.Ag
 		controllerContext.EventRecorder,
 		addonName)
 
-	return addonfactory.NewAgentAddonFactory(addonName, FS, "manifests/managedcluster-chart").
+	return addonfactory.NewAgentAddonFactory(addonName, FS, "manifests/managedclusterchart").
 		WithGetValuesFuncs(getValues, addonfactory.GetValuesFromAddonAnnotation).
 		WithAgentRegistrationOption(registrationOption).
 		BuildHelmAgentAddon()
