+++ pkg/addon/configpolicy/manifests/hubpermissions/role.yaml
