+++ pkg/addon/configpolicy/manifests/hubpermissions/rolebinding.yaml
