+++ pkg/addon/configpolicy/manifests/managedclusterchart/Chart.yaml
