+++ pkg/addon/configpolicy/manifests/managedclusterchart/templates/_helpers.tpl
