+++ pkg/addon/configpolicy/manifests/managedclusterchart/templates/cluster_role.yaml
