+++ pkg/addon/configpolicy/manifests/managedclusterchart/templates/cluster_role_binding.yaml
