+++ pkg/addon/configpolicy/manifests/managedclusterchart/templates/deployment.yaml
