+++ pkg/addon/configpolicy/manifests/managedclusterchart/templates/leader_election_role.yaml
