+++ pkg/addon/configpolicy/manifests/managedclusterchart/templates/service_account.yaml
