+++ pkg/addon/configpolicy/manifests/managedclusterchart/values.yaml
