+++ pkg/addon/iampolicy/manifests/hubpermissions/role.yaml
