+++ pkg/addon/iampolicy/manifests/managedclusterchart/Chart.yaml
