+++ pkg/addon/iampolicy/manifests/managedclusterchart/templates/_helpers.tpl
