+++ pkg/addon/iampolicy/manifests/managedclusterchart/templates/cluster_role.yaml
