+++ pkg/addon/iampolicy/manifests/managedclusterchart/templates/cluster_role_binding.yaml
