+++ pkg/addon/iampolicy/manifests/managedclusterchart/templates/deployment.yaml
