+++ pkg/addon/iampolicy/manifests/managedclusterchart/templates/leader_election_role.yaml
