+++ pkg/addon/iampolicy/manifests/managedclusterchart/templates/service_account.yaml
