+++ pkg/addon/iampolicy/manifests/managedclusterchart/values.yaml
