+++ pkg/addon/policyframework/manifests/hubpermissions/role.yaml
