+++ pkg/addon/policyframework/manifests/managedclusterchart/Chart.yaml
