+++ pkg/addon/policyframework/manifests/managedclusterchart/templates/_helpers.tpl
