+++ pkg/addon/policyframework/manifests/managedclusterchart/templates/cluster_role.yaml
