+++ pkg/addon/policyframework/manifests/managedclusterchart/templates/cluster_role_binding.yaml
