+++ pkg/addon/policyframework/manifests/managedclusterchart/templates/deployment.yaml
