+++ pkg/addon/policyframework/manifests/managedclusterchart/templates/leader_election_role.yaml
