+++ pkg/addon/policyframework/manifests/managedclusterchart/values.yaml
